FGP-1031: moved wmp create offer code to grant-types folder
