+++ src/api/agreement/controllers/get-agreement.controller.js
@@ -1,6 +1,6 @@
 import { statusCodes } from '#~/api/common/constants/status-codes.js'
 import { calculatePaymentsBasedOnParcelsWithActions } from '#~/api/adapter/land-grants-adapter.js'
-import { isWmpAgreement } from '#~/api/agreement/helpers/wmp-payload-mapper.js'
+import { isWmpAgreement } from '#~/api/agreement/helpers/grant-types/wmp/wmp-payload-mapper.js'
 import Boom from '@hapi/boom'
 
 /**
